Cleanup of corpse code
Improved the generation of corpse cloth packets by just checking the
equipped items. (Should speed up crowded areas with many corpses which
have items inside)

The corpse clothing will be sent again to nearby chars if items are
added.

Ensured that the equipped layer is reloaded if the shard restarts.