Documentation: syncing core changex XML with TXT, new automatic tool
The sync now goes back to POL095 era... after that, the format is
completely different and would require writing a 2nd dedicated
parser: I think it's not worth it.

The automatic parser is buildcorechanges.py and is located in a new
"doctools" folder. A real dev may move that folder later where he
will think it is more appropriate.

Parser tested only on Linux, but I wrote it to also work on Windows.