Added: Item insurance support, new Item.insured property (0/1)
An insured item behaves on death like a newbie item, except that
the insured flag gets reset back to 0 after every death. If the
item for some reason (e.g. when backpack is full) will get dropped
instead, the insured flag will be reset anyway. Insured items are
not stackable with normal items but they can still be sold to
vendors.
Hint: this feature can also be used in conjunction with CanDie
sys hook to determine via script which items a player will drop
on death; since the CanDie hook is called exactly one moment
before death you can use it to insure the items that you don't
want to fall on the corpse.